go modules: lock dependencies with vendor folder
We want to ease Kuzzle SDK Go installation and avoid GOPATH and
dependency configuration errors. Also, as part of our effort to
integrate this SDK in Buildroot it is necessary to provide all
dependencies within the same archive.

Go includes support[1] for local copies of external dependencies
in a vendor folder.

Add a vendor directory with all the required dependencies. It has been
generated with the govendor tool[2]. It will be necessary to update and
commit these dependencies regularly. This is possible with "govendor
fetch".

[1] https://golang.org/cmd/go/#hdr-Vendor_Directories
[2] https://github.com/kardianos/govendor